Change m_subpopulation_time_series to int

+++ cpp/tests/test_abm_location.cpp
@@ -24,10 +24,10 @@ TEST(TestLocation, init)
     auto location = mio::abm::Location(mio::abm::LocationType::School, 0);
     for (mio::abm::InfectionState i = mio::abm::InfectionState(0); i < mio::abm::InfectionState::Count;
          i                          = mio::abm::InfectionState(size_t(i) + 1)) {
-        ASSERT_EQ(location.get_subpopulation(i), (unsigned)0);
+        ASSERT_EQ(location.get_subpopulation(i), 0);
     }
     ASSERT_EQ(print_wrap(location.get_time_series_subpopulations().get_last_value()),
-              print_wrap(mio::TimeSeries<unsigned>::Vector::Zero((size_t)mio::abm::InfectionState::Count)));
+              print_wrap(mio::TimeSeries<int>::Vector::Zero((size_t)mio::abm::InfectionState::Count)));
 }
 
 TEST(TestLocation, initCell)
